migrated data collection functions to separate python file
